fix: extract shared mock data package
