--- apps/donor/lib/mock-data/users.ts
@@ -1,935 +1,0 @@
-/**
- * Mock Users Data
- *
- * Comprehensive mock data for all user types: missionaries, donors, and staff.
- * All data is realistic, "lived-in", and ready for use across the application.
- */
-
-import type { Missionary, Donor, StaffMember } from "./types";
-
-const AVATAR_BASE = "https://images.unsplash.com/photo-";
-const COVER_BASE = "https://images.unsplash.com/photo-";
-
-export const MISSIONARIES: Missionary[] = [
-  {
-    id: "miss-001",
-    email: "miller.family@givehope.org",
-    firstName: "David",
-    lastName: "Miller",
-    avatarUrl: `${AVATAR_BASE}1507003211169-0a1dd7228f2d?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1488521787991-ed7bbaae773c?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m1-1",
-        type: "mobile",
-        number: "+66 81 234 5678",
-        isPrimary: true,
-      },
-      {
-        id: "ph-m1-2",
-        type: "work",
-        number: "+66 53 123 456",
-        isPrimary: false,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m1-1",
-        type: "home",
-        street1: "123 Nimman Road",
-        street2: "Soi 9",
-        city: "Chiang Mai",
-        state: "Chiang Mai Province",
-        postalCode: "50200",
-        country: "Thailand",
-        isPrimary: true,
-      },
-      {
-        id: "addr-m1-2",
-        type: "mailing",
-        street1: "PO Box 4521",
-        city: "Denver",
-        state: "CO",
-        postalCode: "80201",
-        country: "USA",
-        isPrimary: false,
-      },
-    ],
-    tags: ["Education", "Southeast Asia", "Church Planting", "Youth Ministry"],
-    notes:
-      "The Miller family has been serving in Thailand since 2018. Strong supporters of local school programs. Annual furlough typically in June-August.",
-    createdAt: "2018-03-15T00:00:00Z",
-    updatedAt: "2024-12-15T08:30:00Z",
-    role: "missionary",
-    title: "The Miller Family",
-    location: "Chiang Mai, Thailand",
-    region: "Southeast Asia",
-    timezone: "Asia/Bangkok",
-    status: "active",
-    healthStatus: "healthy",
-    healthSignals: {
-      emotional: 85,
-      spiritual: 90,
-      physical: 80,
-      financial: 76,
-    },
-    ministryFocus: "Education & Community Development",
-    bio: "The Miller family is dedicated to providing educational opportunities and community development in rural villages of Northern Thailand. Through partnerships with local schools and churches, they have helped over 500 children access quality education.",
-    monthlyGoal: 6000,
-    currentSupport: 4560,
-    donorCount: 142,
-    startDate: "2018-03-15",
-    lastCheckIn: "2024-12-18T09:00:00Z",
-    careGaps: [],
-    projectImage: `${COVER_BASE}1488521787991-ed7bbaae773c?w=800&h=600&fit=crop`,
-  },
-  {
-    id: "miss-002",
-    email: "sarah.smith@givehope.org",
-    firstName: "Sarah",
-    lastName: "Smith",
-    avatarUrl: `${AVATAR_BASE}1494790108377-be9c29b29330?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1519494026892-80bbd2d6fd0d?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m2-1",
-        type: "mobile",
-        number: "+254 712 345 678",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m2-1",
-        type: "home",
-        street1: "Karen Estate, House 45",
-        city: "Nairobi",
-        state: "Nairobi County",
-        postalCode: "00100",
-        country: "Kenya",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Healthcare", "Africa", "Medical Mission", "Women's Health"],
-    notes:
-      "Dr. Smith is a board-certified family physician. Currently dealing with customs delays for medical supplies. Needs prayer support.",
-    createdAt: "2019-06-01T00:00:00Z",
-    updatedAt: "2024-12-19T14:00:00Z",
-    role: "missionary",
-    title: "Dr. Sarah Smith",
-    location: "Nairobi, Kenya",
-    region: "East Africa",
-    timezone: "Africa/Nairobi",
-    status: "active",
-    healthStatus: "at_risk",
-    healthSignals: {
-      emotional: 65,
-      spiritual: 75,
-      physical: 70,
-      financial: 85,
-    },
-    ministryFocus: "Medical Mission & Women's Health",
-    bio: "Dr. Sarah Smith provides essential healthcare services to underserved communities in Nairobi. Her clinic serves over 200 patients weekly, focusing on maternal health and preventable diseases.",
-    monthlyGoal: 8000,
-    currentSupport: 6800,
-    donorCount: 89,
-    startDate: "2019-06-01",
-    lastCheckIn: "2024-12-17T11:00:00Z",
-    careGaps: ["Support Flagged - Customs Issues"],
-    projectImage: `${COVER_BASE}1519494026892-80bbd2d6fd0d?w=800&h=600&fit=crop`,
-  },
-  {
-    id: "miss-003",
-    email: "michael.chen@givehope.org",
-    firstName: "Michael",
-    lastName: "Chen",
-    avatarUrl: `${AVATAR_BASE}1472099645785-5658abf4ff4e?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1469571486292-0ba58a3f068b?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m3-1",
-        type: "mobile",
-        number: "+30 694 123 4567",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m3-1",
-        type: "home",
-        street1: "15 Mytilinis Street",
-        city: "Mytilene",
-        state: "Lesbos",
-        postalCode: "81100",
-        country: "Greece",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Humanitarian", "Europe", "Refugee Ministry", "Crisis Response"],
-    notes:
-      "Michael leads refugee response operations. Fluent in Arabic and Greek. Strong organizational skills.",
-    createdAt: "2020-01-15T00:00:00Z",
-    updatedAt: "2024-12-20T10:30:00Z",
-    role: "missionary",
-    title: "Michael Chen",
-    location: "Lesbos, Greece",
-    region: "Europe",
-    timezone: "Europe/Athens",
-    status: "active",
-    healthStatus: "healthy",
-    healthSignals: {
-      emotional: 70,
-      spiritual: 85,
-      physical: 75,
-      financial: 64,
-    },
-    ministryFocus: "Refugee Crisis Response",
-    bio: "Michael coordinates humanitarian aid for refugees arriving on the shores of Lesbos. His team provides essential supplies, medical care, and emotional support during one of the most challenging moments in these families' lives.",
-    monthlyGoal: 5000,
-    currentSupport: 3200,
-    donorCount: 67,
-    startDate: "2020-01-15",
-    lastCheckIn: "2024-12-19T08:00:00Z",
-    careGaps: ["Financial Gap"],
-    projectImage: `${COVER_BASE}1469571486292-0ba58a3f068b?w=800&h=600&fit=crop`,
-  },
-  {
-    id: "miss-004",
-    email: "olivia.martin@givehope.org",
-    firstName: "Olivia",
-    lastName: "Martin",
-    avatarUrl: `${AVATAR_BASE}1438761681033-6461ffad8d80?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1595053826286-2e59efd9ff18?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m4-1",
-        type: "mobile",
-        number: "+51 987 654 321",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m4-1",
-        type: "home",
-        street1: "Av. Arequipa 1250",
-        street2: "Dept 502",
-        city: "Lima",
-        state: "Lima Province",
-        postalCode: "15046",
-        country: "Peru",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Education", "Latin America", "Youth Development", "Arts"],
-    notes:
-      "Olivia focuses on arts-based education. Currently facing financial stress due to inflation. Needs additional pastoral care.",
-    createdAt: "2021-08-01T00:00:00Z",
-    updatedAt: "2024-12-18T16:00:00Z",
-    role: "missionary",
-    title: "Olivia Martin",
-    location: "Lima, Peru",
-    region: "Latin America",
-    timezone: "America/Lima",
-    status: "active",
-    healthStatus: "at_risk",
-    healthSignals: {
-      emotional: 45,
-      spiritual: 50,
-      physical: 60,
-      financial: 40,
-    },
-    ministryFocus: "Arts Education & Youth Development",
-    bio: "Olivia uses creative arts to reach at-risk youth in Lima's urban neighborhoods. Her programs provide safe spaces for artistic expression while building life skills and community connections.",
-    monthlyGoal: 4500,
-    currentSupport: 2700,
-    donorCount: 45,
-    startDate: "2021-08-01",
-    lastCheckIn: "2024-12-15T14:00:00Z",
-    careGaps: ["Financial Gap", "Support Flagged", "Needs Rest"],
-    projectImage: `${COVER_BASE}1595053826286-2e59efd9ff18?w=800&h=600&fit=crop`,
-  },
-  {
-    id: "miss-005",
-    email: "william.kim@givehope.org",
-    firstName: "William",
-    lastName: "Kim",
-    avatarUrl: `${AVATAR_BASE}1500648767791-00dcc994a43e?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1509099836639-18ba1795216d?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m5-1",
-        type: "mobile",
-        number: "+82 10 9876 5432",
-        isPrimary: true,
-      },
-      {
-        id: "ph-m5-2",
-        type: "work",
-        number: "+82 2 1234 5678",
-        isPrimary: false,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m5-1",
-        type: "home",
-        street1: "234 Gangnam-daero",
-        street2: "Gangnam-gu",
-        city: "Seoul",
-        state: "Seoul",
-        postalCode: "06143",
-        country: "South Korea",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Technology", "Asia", "Digital Ministry", "Training"],
-    notes:
-      "William provides tech support and training for missionaries across Asia. Strong technical background. Good check-in frequency.",
-    createdAt: "2017-11-01T00:00:00Z",
-    updatedAt: "2024-12-20T03:00:00Z",
-    role: "missionary",
-    title: "William Kim",
-    location: "Seoul, South Korea",
-    region: "East Asia",
-    timezone: "Asia/Seoul",
-    status: "active",
-    healthStatus: "healthy",
-    healthSignals: {
-      emotional: 90,
-      spiritual: 85,
-      physical: 95,
-      financial: 90,
-    },
-    ministryFocus: "Technology & Digital Ministry",
-    bio: "William equips missionaries across Asia with technology tools and training. His digital ministry platform has helped over 50 mission organizations streamline their operations and reach more people effectively.",
-    monthlyGoal: 5500,
-    currentSupport: 5200,
-    donorCount: 78,
-    startDate: "2017-11-01",
-    lastCheckIn: "2024-12-19T22:00:00Z",
-    careGaps: [],
-    projectImage: `${COVER_BASE}1509099836639-18ba1795216d?w=800&h=600&fit=crop`,
-  },
-  {
-    id: "miss-006",
-    email: "jackson.lee@givehope.org",
-    firstName: "Jackson",
-    lastName: "Lee",
-    avatarUrl: `${AVATAR_BASE}1519345182560-3f2917c472ef?w=256&h=256&fit=crop&crop=face`,
-    coverImageUrl: `${COVER_BASE}1532629345422-7515f3d16bb6?w=1200&h=600&fit=crop`,
-    phones: [
-      {
-        id: "ph-m6-1",
-        type: "mobile",
-        number: "+254 722 987 654",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-m6-1",
-        type: "home",
-        street1: "Westlands, Plot 45",
-        city: "Nairobi",
-        state: "Nairobi County",
-        postalCode: "00100",
-        country: "Kenya",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Healthcare", "Africa", "Training", "Community Health"],
-    notes:
-      "Jackson trains community health workers. Works closely with Dr. Sarah Smith. Strong local church connections.",
-    createdAt: "2016-04-01T00:00:00Z",
-    updatedAt: "2024-12-14T09:00:00Z",
-    role: "missionary",
-    title: "Jackson Lee",
-    location: "Nairobi, Kenya",
-    region: "East Africa",
-    timezone: "Africa/Nairobi",
-    status: "active",
-    healthStatus: "healthy",
-    healthSignals: {
-      emotional: 70,
-      spiritual: 80,
-      physical: 65,
-      financial: 85,
-    },
-    ministryFocus: "Community Health Training",
-    bio: "Jackson trains local community health workers to provide basic healthcare services in remote villages. His program has trained over 200 health workers who now serve communities with limited access to medical facilities.",
-    monthlyGoal: 4000,
-    currentSupport: 3600,
-    donorCount: 52,
-    startDate: "2016-04-01",
-    lastCheckIn: "2024-12-13T10:00:00Z",
-    careGaps: [],
-    projectImage: `${COVER_BASE}1532629345422-7515f3d16bb6?w=800&h=600&fit=crop`,
-  },
-];
-
-export const DONORS: Donor[] = [
-  {
-    id: "donor-001",
-    email: "john.anderson@email.com",
-    firstName: "John",
-    lastName: "Anderson",
-    avatarUrl: `${AVATAR_BASE}1560250097-0b93528c311a?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d1-1",
-        type: "mobile",
-        number: "+1 (555) 123-4567",
-        isPrimary: true,
-      },
-      {
-        id: "ph-d1-2",
-        type: "home",
-        number: "+1 (555) 987-6543",
-        isPrimary: false,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d1-1",
-        type: "home",
-        street1: "1234 Oak Street",
-        street2: "Apt 5B",
-        city: "Denver",
-        state: "CO",
-        postalCode: "80202",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Major Donor", "Monthly Giver", "Events"],
-    notes:
-      "Long-time supporter of the Miller family. Prefers email communication. Interested in education initiatives.",
-    createdAt: "2019-01-15T00:00:00Z",
-    updatedAt: "2024-12-20T10:00:00Z",
-    role: "donor",
-    stage: "major",
-    company: "Anderson Consulting",
-    jobTitle: "CEO",
-    totalGiven: 45000,
-    totalGivenYTD: 12500,
-    firstGiftDate: "2019-01-20",
-    lastGiftDate: "2024-12-15",
-    lastGiftAmount: 1000,
-    averageGift: 750,
-    largestGift: 5000,
-    giftCount: 60,
-    assignedTo: "staff-001",
-    preferredContact: "email",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: true,
-    },
-  },
-  {
-    id: "donor-002",
-    email: "alice.johnson@techfoundations.org",
-    firstName: "Alice",
-    lastName: "Johnson",
-    avatarUrl: `${AVATAR_BASE}1494790108377-be9c29b29330?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d2-1",
-        type: "work",
-        number: "+1 (415) 555-0101",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d2-1",
-        type: "work",
-        street1: "500 Market Street",
-        street2: "Suite 2000",
-        city: "San Francisco",
-        state: "CA",
-        postalCode: "94102",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Corporate", "High Value", "Water Projects"],
-    notes:
-      "Director of Giving at TechFoundations. Interested in sustainable water projects. Q4 strategy meeting scheduled.",
-    createdAt: "2022-03-01T00:00:00Z",
-    updatedAt: "2024-12-18T14:00:00Z",
-    role: "donor",
-    stage: "major",
-    company: "TechFoundations",
-    jobTitle: "Director of Giving",
-    totalGiven: 75000,
-    totalGivenYTD: 25000,
-    firstGiftDate: "2022-03-15",
-    lastGiftDate: "2024-11-15",
-    lastGiftAmount: 25000,
-    averageGift: 12500,
-    largestGift: 25000,
-    giftCount: 6,
-    assignedTo: "staff-001",
-    preferredContact: "email",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: false,
-    },
-  },
-  {
-    id: "donor-003",
-    email: "bob.smith@globalventures.com",
-    firstName: "Bob",
-    lastName: "Smith",
-    avatarUrl: `${AVATAR_BASE}1500648767791-00dcc994a43e?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d3-1",
-        type: "mobile",
-        number: "+1 (212) 555-0102",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d3-1",
-        type: "work",
-        street1: "350 Fifth Avenue",
-        city: "New York",
-        state: "NY",
-        postalCode: "10118",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Corporate", "Education"],
-    notes:
-      "CEO of Global Ventures. Long-time supporter of educational initiatives in South America.",
-    createdAt: "2020-06-01T00:00:00Z",
-    updatedAt: "2024-12-19T09:00:00Z",
-    role: "donor",
-    stage: "active",
-    company: "Global Ventures",
-    jobTitle: "CEO",
-    totalGiven: 24000,
-    totalGivenYTD: 12000,
-    firstGiftDate: "2020-06-15",
-    lastGiftDate: "2024-12-01",
-    lastGiftAmount: 1000,
-    averageGift: 1000,
-    largestGift: 2000,
-    giftCount: 24,
-    assignedTo: "staff-002",
-    preferredContact: "phone",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: true,
-    },
-  },
-  {
-    id: "donor-004",
-    email: "sarah.connor@email.com",
-    firstName: "Sarah",
-    lastName: "Connor",
-    avatarUrl: `${AVATAR_BASE}1438761681033-6461ffad8d80?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d4-1",
-        type: "mobile",
-        number: "+1 (720) 555-0103",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d4-1",
-        type: "home",
-        street1: "789 Mountain View Drive",
-        city: "Boulder",
-        state: "CO",
-        postalCode: "80302",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Monthly Giver", "Healthcare"],
-    notes:
-      "Passionate about healthcare missions. Supports Dr. Sarah Smith's clinic.",
-    createdAt: "2021-02-01T00:00:00Z",
-    updatedAt: "2024-12-20T08:00:00Z",
-    role: "donor",
-    stage: "active",
-    totalGiven: 7200,
-    totalGivenYTD: 3600,
-    firstGiftDate: "2021-02-15",
-    lastGiftDate: "2024-12-01",
-    lastGiftAmount: 300,
-    averageGift: 300,
-    largestGift: 500,
-    giftCount: 24,
-    preferredContact: "email",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: false,
-      prayerRequests: true,
-    },
-  },
-  {
-    id: "donor-005",
-    email: "mike.wilson@email.com",
-    firstName: "Mike",
-    lastName: "Wilson",
-    avatarUrl: `${AVATAR_BASE}1472099645785-5658abf4ff4e?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d5-1",
-        type: "mobile",
-        number: "+1 (303) 555-0104",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d5-1",
-        type: "home",
-        street1: "456 Pine Street",
-        city: "Denver",
-        state: "CO",
-        postalCode: "80203",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["New Donor", "Youth Ministry"],
-    notes: "Recently started supporting. Interested in youth programs.",
-    createdAt: "2024-09-01T00:00:00Z",
-    updatedAt: "2024-12-15T11:00:00Z",
-    role: "donor",
-    stage: "new",
-    totalGiven: 750,
-    totalGivenYTD: 750,
-    firstGiftDate: "2024-09-15",
-    lastGiftDate: "2024-12-01",
-    lastGiftAmount: 250,
-    averageGift: 250,
-    largestGift: 250,
-    giftCount: 3,
-    preferredContact: "email",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: false,
-    },
-  },
-  {
-    id: "donor-006",
-    email: "jennifer.davis@email.com",
-    firstName: "Jennifer",
-    lastName: "Davis",
-    avatarUrl: `${AVATAR_BASE}1544005313-94ddf0286df2?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d6-1",
-        type: "home",
-        number: "+1 (512) 555-0105",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d6-1",
-        type: "home",
-        street1: "2100 Congress Avenue",
-        city: "Austin",
-        state: "TX",
-        postalCode: "78701",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Lapsed", "Needs Outreach"],
-    notes:
-      "Hasn't given in 8 months. Previously supported Michael Chen's refugee ministry.",
-    createdAt: "2020-01-15T00:00:00Z",
-    updatedAt: "2024-04-01T10:00:00Z",
-    role: "donor",
-    stage: "lapsed",
-    totalGiven: 3600,
-    totalGivenYTD: 0,
-    firstGiftDate: "2020-01-20",
-    lastGiftDate: "2024-04-01",
-    lastGiftAmount: 100,
-    averageGift: 100,
-    largestGift: 200,
-    giftCount: 36,
-    preferredContact: "mail",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: false,
-      prayerRequests: false,
-    },
-  },
-  {
-    id: "donor-007",
-    email: "grace.community@church.org",
-    firstName: "Church of Grace",
-    lastName: "(Organization)",
-    avatarUrl: `${AVATAR_BASE}1478147112227-4b24f2d55906?w=256&h=256&fit=crop`,
-    phones: [
-      {
-        id: "ph-d7-1",
-        type: "work",
-        number: "+1 (214) 555-0106",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d7-1",
-        type: "work",
-        street1: "1500 Main Street",
-        city: "Dallas",
-        state: "TX",
-        postalCode: "75201",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Church Partner", "Monthly Giver", "Major Donor"],
-    notes:
-      "Church partner supporting multiple missionaries. Pledge is currently past due - follow up needed.",
-    createdAt: "2018-01-01T00:00:00Z",
-    updatedAt: "2024-12-01T10:00:00Z",
-    role: "donor",
-    stage: "active",
-    company: "Church of Grace",
-    totalGiven: 120000,
-    totalGivenYTD: 18000,
-    firstGiftDate: "2018-01-15",
-    lastGiftDate: "2024-10-01",
-    lastGiftAmount: 2000,
-    averageGift: 2000,
-    largestGift: 5000,
-    giftCount: 60,
-    assignedTo: "staff-001",
-    preferredContact: "email",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: true,
-    },
-  },
-  {
-    id: "donor-008",
-    email: "robert.martinez@email.com",
-    firstName: "Robert",
-    lastName: "Martinez",
-    avatarUrl: `${AVATAR_BASE}1507591064344-4c6ce005b128?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-d8-1",
-        type: "mobile",
-        number: "+1 (305) 555-0107",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-d8-1",
-        type: "home",
-        street1: "800 Brickell Avenue",
-        street2: "Unit 3401",
-        city: "Miami",
-        state: "FL",
-        postalCode: "33131",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Prospect", "High Potential", "Latin America"],
-    notes:
-      "Met at annual gala. Very interested in Latin American missions. Schedule follow-up call.",
-    createdAt: "2024-11-15T00:00:00Z",
-    updatedAt: "2024-12-10T15:00:00Z",
-    role: "donor",
-    stage: "prospect",
-    company: "Martinez Holdings",
-    jobTitle: "Managing Partner",
-    totalGiven: 0,
-    totalGivenYTD: 0,
-    firstGiftDate: "",
-    lastGiftDate: "",
-    lastGiftAmount: 0,
-    averageGift: 0,
-    largestGift: 0,
-    giftCount: 0,
-    assignedTo: "staff-002",
-    preferredContact: "phone",
-    communicationPreferences: {
-      newsletter: true,
-      taxReceipts: true,
-      eventInvites: true,
-      prayerRequests: false,
-    },
-  },
-];
-
-export const STAFF: StaffMember[] = [
-  {
-    id: "staff-001",
-    email: "admin@givehope.org",
-    firstName: "Emily",
-    lastName: "Thompson",
-    avatarUrl: `${AVATAR_BASE}1580489944761-15a19d654956?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-s1-1",
-        type: "work",
-        number: "+1 (555) 100-0001",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-s1-1",
-        type: "work",
-        street1: "100 Mission Way",
-        city: "Colorado Springs",
-        state: "CO",
-        postalCode: "80903",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Leadership", "Operations"],
-    notes: "Executive Director. Primary contact for major donors.",
-    createdAt: "2015-01-01T00:00:00Z",
-    updatedAt: "2024-12-20T08:00:00Z",
-    role: "admin",
-    title: "Executive Director",
-    department: "Leadership",
-    hireDate: "2015-01-01",
-  },
-  {
-    id: "staff-002",
-    email: "care@givehope.org",
-    firstName: "David",
-    lastName: "Rodriguez",
-    avatarUrl: `${AVATAR_BASE}1506794778202-cad84cf45f1d?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-s2-1",
-        type: "work",
-        number: "+1 (555) 100-0002",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-s2-1",
-        type: "work",
-        street1: "100 Mission Way",
-        city: "Colorado Springs",
-        state: "CO",
-        postalCode: "80903",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Missionary Care", "Pastoral"],
-    notes: "Missionary Care Lead. Handles all care check-ins and support.",
-    createdAt: "2018-06-01T00:00:00Z",
-    updatedAt: "2024-12-19T14:00:00Z",
-    role: "staff",
-    title: "Missionary Care Lead",
-    department: "Missionary Care",
-    hireDate: "2018-06-01",
-  },
-  {
-    id: "staff-003",
-    email: "finance@givehope.org",
-    firstName: "Rachel",
-    lastName: "Kim",
-    avatarUrl: `${AVATAR_BASE}1573496359142-b8d87734a5a2?w=256&h=256&fit=crop&crop=face`,
-    phones: [
-      {
-        id: "ph-s3-1",
-        type: "work",
-        number: "+1 (555) 100-0003",
-        isPrimary: true,
-      },
-    ],
-    addresses: [
-      {
-        id: "addr-s3-1",
-        type: "work",
-        street1: "100 Mission Way",
-        city: "Colorado Springs",
-        state: "CO",
-        postalCode: "80903",
-        country: "USA",
-        isPrimary: true,
-      },
-    ],
-    tags: ["Finance", "Compliance"],
-    notes:
-      "Finance Director. Manages all financial reporting and donor receipts.",
-    createdAt: "2019-03-15T00:00:00Z",
-    updatedAt: "2024-12-18T16:00:00Z",
-    role: "staff",
-    title: "Finance Director",
-    department: "Finance",
-    hireDate: "2019-03-15",
-  },
-];
-
-export function getMissionaryById(id: string): Missionary | undefined {
-  return MISSIONARIES.find((m) => m.id === id);
-}
-
-export function getDonorById(id: string): Donor | undefined {
-  return DONORS.find((d) => d.id === id);
-}
-
-export function getStaffById(id: string): StaffMember | undefined {
-  return STAFF.find((s) => s.id === id);
-}
-
-export function getMissionariesByRegion(region: string): Missionary[] {
-  return MISSIONARIES.filter((m) => m.region === region);
-}
-
-export function getDonorsByStage(stage: string): Donor[] {
-  return DONORS.filter((d) => d.stage === stage);
-}
-
-export function getActiveMissionaries(): Missionary[] {
-  return MISSIONARIES.filter((m) => m.status === "active");
-}
-
-export function getMissionariesAtRisk(): Missionary[] {
-  return MISSIONARIES.filter(
-    (m) => m.healthStatus === "at_risk" || m.healthStatus === "critical",
-  );
-}
-
-export function getDonorFullName(donor: Donor): string {
-  return `${donor.firstName} ${donor.lastName}`.trim();
-}
-
-export function getMissionaryFullName(missionary: Missionary): string {
-  return (
-    missionary.title || `${missionary.firstName} ${missionary.lastName}`.trim()
-  );
-}
