+++ apps/donor/next.config.ts
@@ -38,6 +38,7 @@ const nextConfig: NextConfig = {
     "@asym/ui",
     "@asym/database",
     "@asym/lib",
+    "@asym/mock-data",
     "@asym/config",
     "@asym/auth",
     "@asym/email",
