+++ apps/donor/package.json
@@ -25,6 +25,7 @@
     "@asym/env": "workspace:*",
     "@asym/graphql": "workspace:*",
     "@asym/lib": "workspace:*",
+    "@asym/mock-data": "workspace:*",
     "@asym/ui": "workspace:*",
     "@base-ui/react": "1.5.0",
     "@openpolicy/react": "0.0.17",
