+++ apps/missionary/package.json
@@ -19,6 +19,7 @@
     "@asym/email": "workspace:*",
     "@asym/env": "workspace:*",
     "@asym/lib": "workspace:*",
+    "@asym/mock-data": "workspace:*",
     "@asym/missionary": "workspace:*",
     "@asym/ui": "workspace:*",
     "@base-ui/react": "1.5.0",
