+++ packages/mock-data/eslint.config.mjs
@@ -1,0 +1,5 @@
+import { libraryConfig } from "@asym/eslint-config/library.mjs";
+
+export { libraryConfig };
+
+export default libraryConfig;
