+++ packages/mock-data/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@asym/mock-data",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./index.ts"
+  },
+  "scripts": {
+    "build": "tsc --noEmit",
+    "lint": "eslint .",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@asym/eslint-config": "workspace:*",
+    "@asym/typescript-config": "workspace:*"
+  }
+}
